src/utils.py
def extract_text_from_image(image):
    import pytesseract
    from PIL import Image

    # Use pytesseract to do OCR on the image
    text = pytesseract.image_to_string(image, lang='jpn')
    return text

def preprocess_image(image):
    from PIL import ImageFilter

    # Apply some preprocessing to the image (e.g., convert to grayscale and apply a filter)
    processed_image = image.convert('L').filter(ImageFilter.SHARPEN)
    return processed_image

def save_image(image, file_path):
    # Save the image to the specified file path
    image.save(file_path)