requirements.txt
Pillow
pytesseract
googletrans==4.0.0-rc1
opencv-python
numpy